changement structure fichier
