+++ argentbank-app/src/App.test.js
@@ -5,11 +5,11 @@ import { store } from './app/store';
 import App from './App';
 
 test('renders learn react link', () => {
-  const { getByText } = render(
-    <Provider store={store}>
-      <App />
-    </Provider>
-  );
+	const { getByText } = render(
+		<Provider store={store}>
+			<App />
+		</Provider>
+	);
 
-  expect(getByText(/learn/i)).toBeInTheDocument();
+	expect.screen.getByText(/learn/i).toBeInTheDocument();
 });
